Move render-related code to renderer package
